+++ Libplanet/Blockchain/BlockChain.cs
@@ -290,9 +290,27 @@ namespace Libplanet.Blockchain
                 transactions.Select(tx => tx.Id).ToImmutableHashSet());
         }
 
-        public long GetNonce(Address address)
+        public long GetNonce(Address address, bool includeStage = false)
         {
-            return Store.GetTxNonce(Id.ToString(), address);
+            long nonce = Store.GetTxNonce(Id.ToString(), address);
+
+            if (includeStage)
+            {
+                IEnumerable<Transaction<T>> stagedTxs = Store
+                    .IterateStagedTransactionIds()
+                    .Select(Store.GetTransaction<T>)
+                    .Where(tx => tx.Signer.Equals(address));
+
+                foreach (Transaction<T> tx in stagedTxs)
+                {
+                    if (nonce <= tx.Nonce)
+                    {
+                        nonce = tx.Nonce + 1;
+                    }
+                }
+            }
+
+            return nonce;
         }
 
         public Block<T> MineBlock(